pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Pau <plabartabajo@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
bytewax = "^0.15.1"
websocket-client = "^1.5.1"
pytz = "^2022.7.1"
numpy = "^1.24.2"
streamlit = "^1.20.0"
bokeh = "^2.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
